app/requirements.txt
strands-agents
strands-agents-tools
streamlit
fastapi
uvicorn
requests
python-dotenv